DeepLearning/app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart
pillow
numpy
tensorflow==2.20.0

